- Rename from esbuild_inline to esbuild - Change indentation to 2 spaces
