+++ examples/esbuild/package.json
