+++ examples/esbuild/src/fib.js
